time(02:06:05): Inheritance => Reusing methods and attributes
Reusing methods and attributes from the parent-class/superclass on
child-classes/subclass. In this case all methods that are defined on `class
Item` could be used in `class Phone`, while there are also specific methods that
are defined only on `class Phone` won't interfene with the parent-class.

Remember: Certain methods or attributes could be defined in the subclass to
override the defined methods and attributes in the superclass