time(02:12:21): Polymorphism => Ability to take different kind input
It is the ability to take different inputs and displaying the output
accordingly. It works hand in hand with abstraction through Interface (point of
contact) and not directly modifying the Implementation.
In this case every REPL operation on the subclass would still make methods in
the superclass functional. It is proven by creating new subclass in
`keyboard.py` with it's own specific configurations.